+++ tools/schema/group_price/create_group_price.rb
@@ -3,7 +3,7 @@ require_relative '../tools_helper'
 cql = %Q(CREATE TABLE group_prices
         (
            sku  bigint Primary Key,
-            standardPrice  double,
-            prices map<text,double>,
-            discount double))
+            standardPrice  decimal,
+            prices map<text,decimal>,
+            discount decimal))
 execute_lazy cql, []
